+++ src/pybamm/simulation.py
@@ -610,6 +610,12 @@ class Simulation:
 
         elif self.operating_mode == "with experiment":
             callbacks.on_experiment_start(logs)
+
+            if isinstance(inputs, list):
+                raise pybamm.SolverError(
+                    "Solving with a list of input sets is not supported with experiments."
+                )
+
             self.build_for_experiment(
                 initial_soc=initial_soc,
                 direction=direction,